requirements.txt
streamlit==1.39.0
pandas==2.2.3
matplotlib==3.9.2
python-dotenv== 1.0.0
ipykernel==6.25.2
seaborn